TW-503: Buy section. Binance Connect. Refactor
